Move dss.util.checkout to dss.storage.checkout
